+++ sdk/storage/azure-storage-file-share/azure/storage/fileshare/_share_client.py
@@ -130,7 +130,8 @@ class ShareClient(StorageAccountHostsMixin): # pylint: disable=too-many-public-m
         self.share_name = share_name
         self._query_str, credential = self._format_query_string(
             sas_token, credential, share_snapshot=self.snapshot)
-        super(ShareClient, self).__init__(parsed_url, service='file-share', credential=credential, **kwargs)
+        super(ShareClient, self).__init__(
+            parsed_url, service='file-share', credential=credential, **kwargs)
         self.allow_trailing_dot = kwargs.pop('allow_trailing_dot', None)
         self.allow_source_trailing_dot = kwargs.pop('allow_source_trailing_dot', None)
         self.file_request_intent = token_intent